requirement.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
requests==2.28.1
gitpython==3.1.31
speechrecognition==3.8.1
pillow==9.0.1
markdown==3.4.3
python-dotenv==0.21.1
